move card into src components, rename Card to InfoCard
