+++ src/components/ui/Card.tsx
@@ -1,101 +1,103 @@
-import Link from "next/link";
-import styles from "../../styles/Card.module.css";
-import Image from "next/image";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faExternalLinkAlt } from "@fortawesome/free-solid-svg-icons";
+import * as React from "react";
 
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-  CardFooter,
-} from "@/components/ui/card";
+import { cn } from "../../utils/utils";
 
-type imageLocations = "left" | "right";
-type imageOptions = {
-  src: string;
-  location: imageLocations;
-  alt: string;
-};
-interface CardProps {
-  header: React.ReactNode;
-  isPinned?: boolean;
-  description?: string;
-  content?: React.ReactNode;
-  link?: string;
-  linkHoverText?: string;
-  imageOptions?: imageOptions;
-  footer?: React.ReactNode;
+function Card({
+  className,
+  size = "default",
+  ...props
+}: React.ComponentProps<"div"> & { size?: "default" | "sm" }) {
+  return (
+    <div
+      data-slot="card"
+      data-size={size}
+      className={cn(
+        "group/card flex flex-col gap-6 overflow-hidden rounded-xl bg-card py-6 text-sm text-card-foreground shadow-xs ring-1 ring-foreground/10 has-[>img:first-child]:pt-0 data-[size=sm]:gap-4 data-[size=sm]:py-4 *:[img:first-child]:rounded-t-xl *:[img:last-child]:rounded-b-xl",
+        className
+      )}
+      {...props}
+    />
+  );
 }
 
-export const InfoCard = ({
-  header: title,
-  isPinned,
-  description,
-  content,
-  link,
-  imageOptions,
-  linkHoverText,
-  footer,
-}: CardProps) => {
-  if (description) {
-    description = description.replaceAll("\n", "--");
-    description = description.replaceAll("--", "\n");
-  }
-
+function CardHeader({ className, ...props }: React.ComponentProps<"div">) {
   return (
-    <Card
-    // className={`${styles.card}`}
-    // style={{ border: "1px solid var(--clr-primary)" }}
-    >
-      <CardHeader>
-        {imageOptions ?
-          imageOptions.location == "left" ?
-            <Image src={imageOptions.src} alt={imageOptions.alt} fill />
-          : null
-        : null}
-
-        <CardTitle>
-          <h2 className="border-bottom">
-            {isPinned ?
-              <span className={styles.pinned}>📌 </span>
-            : null}
-            {link ?
-              <span>
-                <Link
-                  className={styles.link}
-                  href={link}
-                  target="_blank"
-                  rel="noreferrer"
-                  title={linkHoverText || "Link to project"}
-                >
-                  {title} <FontAwesomeIcon icon={faExternalLinkAlt} />
-                </Link>
-              </span>
-            : title}
-          </h2>
-        </CardTitle>
-      </CardHeader>
-
-      <CardDescription>
-        <p>{description}</p>
-      </CardDescription>
-
-      <CardContent>
-        {content ? content : null}
-
-        {imageOptions ?
-          imageOptions.location == "right" ?
-            <Image src={imageOptions.src} alt={imageOptions.alt} />
-          : null
-        : null}
-
-        {footer ?
-          <div className={styles.footer}>{footer}</div>
-        : null}
-      </CardContent>
-    </Card>
+    <div
+      data-slot="card-header"
+      className={cn(
+        "group/card-header @container/card-header grid auto-rows-min items-start gap-1 rounded-t-xl px-6 group-data-[size=sm]/card:px-4 has-data-[slot=card-action]:grid-cols-[1fr_auto] has-data-[slot=card-description]:grid-rows-[auto_auto] [.border-b]:pb-6 group-data-[size=sm]/card:[.border-b]:pb-4",
+        className
+      )}
+      {...props}
+    />
   );
+}
+
+function CardTitle({ className, ...props }: React.ComponentProps<"div">) {
+  return (
+    <div
+      data-slot="card-title"
+      className={cn(
+        "font-heading text-base leading-normal font-medium group-data-[size=sm]/card:text-sm",
+        className
+      )}
+      {...props}
+    />
+  );
+}
+
+function CardDescription({ className, ...props }: React.ComponentProps<"div">) {
+  return (
+    <div
+      data-slot="card-description"
+      className={cn("text-sm text-muted-foreground", className)}
+      {...props}
+    />
+  );
+}
+
+function CardAction({ className, ...props }: React.ComponentProps<"div">) {
+  return (
+    <div
+      data-slot="card-action"
+      className={cn(
+        "col-start-2 row-span-2 row-start-1 self-start justify-self-end",
+        className
+      )}
+      {...props}
+    />
+  );
+}
+
+function CardContent({ className, ...props }: React.ComponentProps<"div">) {
+  return (
+    <div
+      data-slot="card-content"
+      className={cn("px-6 group-data-[size=sm]/card:px-4", className)}
+      {...props}
+    />
+  );
+}
+
+function CardFooter({ className, ...props }: React.ComponentProps<"div">) {
+  return (
+    <div
+      data-slot="card-footer"
+      className={cn(
+        "flex items-center rounded-b-xl px-6 group-data-[size=sm]/card:px-4 [.border-t]:pt-6 group-data-[size=sm]/card:[.border-t]:pt-4",
+        className
+      )}
+      {...props}
+    />
+  );
+}
+
+export {
+  Card,
+  CardHeader,
+  CardFooter,
+  CardTitle,
+  CardAction,
+  CardDescription,
+  CardContent,
 };
